Add Enums and OOP
