+++ src/main/java/NeroException.java
@@ -1,0 +1,6 @@
+public class NeroException extends Exception {
+
+    public NeroException(String errorMessage) {
+        super(errorMessage);
+    }
+}
