+++ src/main/java/Task.java
@@ -1,0 +1,32 @@
+public abstract class Task {
+    protected String description;
+    protected boolean isDone;
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    public abstract String currToPrint();
+    public String getDescription() {
+        return description;
+    }
+
+    public boolean getStatus() {
+        return isDone;
+    }
+
+    public String getStatusIcon() {
+        return (isDone ? "[X]" : "[ ]");
+    }
+
+    public void markAsDone() {
+        isDone = true;
+    }
+
+    public void markAsUndone(){
+        isDone = false;
+    }
+
+
+}
+
