+++ src/main/java/TaskList.java
@@ -1,0 +1,32 @@
+import javax.lang.model.type.ArrayType;
+import java.util.ArrayList;
+
+public class TaskList<Task> {
+    public ArrayList<Task> taskList;
+
+    public TaskList() {
+        this.taskList = new ArrayList<Task>();
+    }
+
+    void addTask(Task task) {
+        this.taskList.add(task);
+    }
+
+    Task get(int index) {
+        return taskList.get(index);
+    }
+
+    void removeTask(int index) throws NeroException {
+        try {
+            this.taskList.remove(index);
+        } catch (IndexOutOfBoundsException e) {
+            throw new NeroException("Please enter a valid index from 0 to " + (this.taskList.size() - 1));
+        }
+    }
+
+    void printTasks() {
+        for (int i = 0; i < taskList.size(); i++) {
+            System.out.println(taskList.get(i));
+        }
+    }
+}
