+++ src/main/java/ToDo.java
@@ -1,0 +1,13 @@
+public class ToDo extends Task {
+    public ToDo(String description) {
+        super(description);
+    }
+
+    public String getTaskIcon() {
+        return "[T]";
+    }
+
+    public String currToPrint() {
+        return this.getTaskIcon() + this.getStatusIcon() + " " + this.getDescription();
+    }
+}
